arrival-app/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
anthropic==0.43.0
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
PyJWT==2.10.1
